requirements.txt
requests
requests-oauthlib
oauthlib
scipy
pandas
shapely
rasterio
xarray
retrying
pip-system-certs
PyJWT
cryptography